+++ compiler/src/iree/compiler/Codegen/Common/BUILD.bazel
@@ -153,6 +153,7 @@ iree_compiler_cc_library(
         "TileAndDistributeToWorkgroupsPass.cpp",
         "TileDispatchUsingForall.cpp",
         "TileDispatchUsingInterface.cpp",
+        "TileInferenceUtils.cpp",
         "TileLargeTensors.cpp",
         "TileSizeSelection.cpp",
         "Transforms.cpp",
@@ -170,6 +171,7 @@ iree_compiler_cc_library(
         "PassUtils.h",
         "Passes.h",
         "TensorDynamicDimAnalysis.h",
+        "TileInferenceUtils.h",
         "TileSizeSelection.h",
         "Transforms.h",
         "UserConfig.h",
