+++ compiler/src/iree/compiler/Codegen/Common/CMakeLists.txt
@@ -74,6 +74,7 @@ iree_cc_library(
     "PassUtils.h"
     "Passes.h"
     "TensorDynamicDimAnalysis.h"
+    "TileInferenceUtils.h"
     "TileSizeSelection.h"
     "Transforms.h"
     "UserConfig.h"
@@ -145,6 +146,7 @@ iree_cc_library(
     "TileAndDistributeToWorkgroupsPass.cpp"
     "TileDispatchUsingForall.cpp"
     "TileDispatchUsingInterface.cpp"
+    "TileInferenceUtils.cpp"
     "TileLargeTensors.cpp"
     "TileSizeSelection.cpp"
     "Transforms.cpp"
