+++ compiler/src/iree/compiler/Codegen/LLVMGPU/test/ROCDL/config_tile_and_fuse.mlir
@@ -194,7 +194,7 @@ module {
 //  CHECK-SAME:     promote_operands = [0, 1]
 //  CHECK-SAME:     reduction = [0, 0, 4]
 //  CHECK-SAME:     thread = [1, 4, 0]
-//  CHECK-SAME:     workgroup = [1, 256, 0]
+//  CHECK-SAME:     workgroup = [1, 0, 0]
 
 //        LATE:  LLVMGPUWarpReduction
 
@@ -215,7 +215,7 @@ module {
 //  CHECK-SAME:   #iree_codegen.translation_info<pipeline = LLVMGPUTileAndFuse workgroup_size = [64, 1, 1] subgroup_size = 64>
 //       CHECK:   linalg.add {{.*}}lowering_config = #iree_gpu.lowering_config
 //  CHECK-SAME:     thread = [1, 4]
-//  CHECK-SAME:     workgroup = [1, 256]
+//  CHECK-SAME:     workgroup = [1, 0]
 
 // -----
 
@@ -576,3 +576,81 @@ func.func @unset_encoding_gpu(%arg0: tensor<10x5x4x8x2x4x16x4xi32>) -> tensor<12
 //       CHECK:   linalg.generic {{.*}}lowering_config = #iree_gpu.lowering_config
 //  CHECK-SAME:     thread = [1, 1, 1, 1, 1, 1, 1, 1]
 //  CHECK-SAME:     workgroup = [1, 1, 0, 0, 0, 0, 0, 0]
+
+// -----
+
+func.func @pack_dynamic_producer(%arg0: tensor<?x?xi8>, %d0: index, %d1: index, %d2: index, %d3: index) -> tensor<?x?x32x32xi8> {
+  %c0_i8 = arith.constant 0 : i8
+  %init0 = tensor.empty(%d0, %d1) : tensor<?x?xi8>
+  %0 = linalg.generic {
+      indexing_maps = [affine_map<(d0, d1) -> (d1, d0)>,
+                       affine_map<(d0, d1) -> (d0, d1)>],
+      iterator_types = ["parallel", "parallel"]}
+      ins(%arg0 : tensor<?x?xi8>) outs(%init0 : tensor<?x?xi8>) {
+  ^bb0(%in: i8, %out: i8):
+    linalg.yield %in : i8
+  } -> tensor<?x?xi8>
+  %init1 = tensor.empty(%d2, %d3) : tensor<?x?x32x32xi8>
+  %pack = tensor.pack %0 padding_value(%c0_i8 : i8)
+      outer_dims_perm = [0, 1] inner_dims_pos = [0, 1] inner_tiles = [32, 32]
+      into %init1 : tensor<?x?xi8> -> tensor<?x?x32x32xi8>
+  return %pack : tensor<?x?x32x32xi8>
+}
+
+// CHECK-LABEL: func.func @pack_dynamic_producer
+//  CHECK-SAME:   #iree_codegen.translation_info<pipeline = LLVMGPUTileAndFuse workgroup_size = [1, 1, 1] subgroup_size = 64>
+//       CHECK:   linalg.generic {{.*}}lowering_config = #iree_gpu.lowering_config
+//  CHECK-SAME:     thread = [1, 1]
+//  CHECK-SAME:     workgroup = [32, 32]
+
+// -----
+
+func.func @pack_full_tile(%arg0: tensor<32x32xi8>) -> tensor<1x1x32x32xi8> {
+  %c0_i8 = arith.constant 0 : i8
+  %init0 = tensor.empty() : tensor<32x32xi8>
+  %0 = linalg.generic {
+      indexing_maps = [affine_map<(d0, d1) -> (d1, d0)>,
+                       affine_map<(d0, d1) -> (d0, d1)>],
+      iterator_types = ["parallel", "parallel"]}
+      ins(%arg0 : tensor<32x32xi8>) outs(%init0 : tensor<32x32xi8>) {
+  ^bb0(%in: i8, %out: i8):
+    linalg.yield %in : i8
+  } -> tensor<32x32xi8>
+  %init1 = tensor.empty() : tensor<1x1x32x32xi8>
+  %pack = tensor.pack %0 padding_value(%c0_i8 : i8)
+      outer_dims_perm = [0, 1] inner_dims_pos = [0, 1] inner_tiles = [32, 32]
+      into %init1 : tensor<32x32xi8> -> tensor<1x1x32x32xi8>
+  return %pack : tensor<1x1x32x32xi8>
+}
+
+// CHECK-LABEL: func.func @pack_full_tile
+//  CHECK-SAME:   #iree_codegen.translation_info<pipeline = LLVMGPUTileAndFuse workgroup_size = [64, 1, 1] subgroup_size = 64>
+//       CHECK:   linalg.generic {{.*}}lowering_config = #iree_gpu.lowering_config
+//  CHECK-SAME:     thread = [1, 4]
+//  CHECK-SAME:     workgroup = [0, 0]
+
+// -----
+
+func.func @pack_dynamic_tile(%arg0: tensor<32x32xi8>, %d0: index, %d1: index, %tile0: index, %tile1: index) -> tensor<?x?x?x?xi8> {
+  %c0_i8 = arith.constant 0 : i8
+  %init0 = tensor.empty() : tensor<32x32xi8>
+  %0 = linalg.generic {
+      indexing_maps = [affine_map<(d0, d1) -> (d1, d0)>,
+                       affine_map<(d0, d1) -> (d0, d1)>],
+      iterator_types = ["parallel", "parallel"]}
+      ins(%arg0 : tensor<32x32xi8>) outs(%init0 : tensor<32x32xi8>) {
+  ^bb0(%in: i8, %out: i8):
+    linalg.yield %in : i8
+  } -> tensor<32x32xi8>
+  %init1 = tensor.empty(%d0, %d1, %tile0, %tile1) : tensor<?x?x?x?xi8>
+  %pack = tensor.pack %0 padding_value(%c0_i8 : i8)
+      outer_dims_perm = [0, 1] inner_dims_pos = [0, 1] inner_tiles = [%tile0, %tile1]
+      into %init1 : tensor<32x32xi8> -> tensor<?x?x?x?xi8>
+  return %pack : tensor<?x?x?x?xi8>
+}
+
+// CHECK-LABEL: func.func @pack_dynamic_tile
+//  CHECK-SAME:   #iree_codegen.translation_info<pipeline = LLVMGPUTileAndFuse workgroup_size = [64, 1, 1] subgroup_size = 64>
+//       CHECK:   linalg.generic {{.*}}lowering_config = #iree_gpu.lowering_config
+//  CHECK-SAME:     thread = [1, 4]
+//  CHECK-SAME:     workgroup = [8, 0]
